+++ modules/lang-painless/src/main/java/org/elasticsearch/painless/node/EUnary.java
@@ -39,6 +39,7 @@ public final class EUnary extends AExpression {
 
     final Operation operation;
     AExpression child;
+    Type promote;
 
     public EUnary(Location location, Operation operation, AExpression child) {
         super(location);
@@ -77,7 +78,7 @@ public final class EUnary extends AExpression {
     void analyzeBWNot(Locals variables) {
         child.analyze(variables);
 
-        Type promote = AnalyzerCaster.promoteNumeric(child.actual, false);
+        promote = AnalyzerCaster.promoteNumeric(child.actual, false);
 
         if (promote == null) {
             throw createError(new ClassCastException("Cannot apply not [~] to type [" + child.actual.name + "]."));
@@ -98,13 +99,17 @@ public final class EUnary extends AExpression {
             }
         }
 
-        actual = promote;
+        if (promote.sort == Sort.DEF && expected != null) {
+            actual = expected;
+        } else {
+            actual = promote;
+        }
     }
 
     void analyzerAdd(Locals variables) {
         child.analyze(variables);
 
-        Type promote = AnalyzerCaster.promoteNumeric(child.actual, true);
+        promote = AnalyzerCaster.promoteNumeric(child.actual, true);
 
         if (promote == null) {
             throw createError(new ClassCastException("Cannot apply positive [+] to type [" + child.actual.name + "]."));
@@ -129,13 +134,17 @@ public final class EUnary extends AExpression {
             }
         }
 
-        actual = promote;
+        if (promote.sort == Sort.DEF && expected != null) {
+            actual = expected;
+        } else {
+            actual = promote;
+        }
     }
 
     void analyzerSub(Locals variables) {
         child.analyze(variables);
 
-        Type promote = AnalyzerCaster.promoteNumeric(child.actual, true);
+        promote = AnalyzerCaster.promoteNumeric(child.actual, true);
 
         if (promote == null) {
             throw createError(new ClassCastException("Cannot apply negative [-] to type [" + child.actual.name + "]."));
@@ -160,7 +169,11 @@ public final class EUnary extends AExpression {
             }
         }
 
-        actual = promote;
+        if (promote.sort == Sort.DEF && expected != null) {
+            actual = expected;
+        } else {
+            actual = promote;
+        }
     }
 
     @Override
@@ -186,14 +199,12 @@ public final class EUnary extends AExpression {
                 child.write(writer);
             }
         } else {
-            org.objectweb.asm.Type type = actual.type;
-            Sort sort = actual.sort;
-
+            Sort sort = promote.sort;
             child.write(writer);
 
             if (operation == Operation.BWNOT) {
                 if (sort == Sort.DEF) {
-                    org.objectweb.asm.Type descriptor = org.objectweb.asm.Type.getMethodType(expected.type, child.actual.type);
+                    org.objectweb.asm.Type descriptor = org.objectweb.asm.Type.getMethodType(actual.type, child.actual.type);
                     writer.invokeDynamic("not", descriptor.getDescriptor(), DEF_BOOTSTRAP_HANDLE, DefBootstrap.UNARY_OPERATOR);
                 } else {
                     if (sort == Sort.INT) {
@@ -204,18 +215,18 @@ public final class EUnary extends AExpression {
                         throw createError(new IllegalStateException("Illegal tree structure."));
                     }
 
-                    writer.math(MethodWriter.XOR, type);
+                    writer.math(MethodWriter.XOR, actual.type);
                 }
             } else if (operation == Operation.SUB) {
                 if (sort == Sort.DEF) {
-                    org.objectweb.asm.Type descriptor = org.objectweb.asm.Type.getMethodType(expected.type, child.actual.type);
+                    org.objectweb.asm.Type descriptor = org.objectweb.asm.Type.getMethodType(actual.type, child.actual.type);
                     writer.invokeDynamic("neg", descriptor.getDescriptor(), DEF_BOOTSTRAP_HANDLE, DefBootstrap.UNARY_OPERATOR);
                 } else {
-                    writer.math(MethodWriter.NEG, type);
+                    writer.math(MethodWriter.NEG, actual.type);
                 }
             } else if (operation == Operation.ADD) {
                 if (sort == Sort.DEF) {
-                    org.objectweb.asm.Type descriptor = org.objectweb.asm.Type.getMethodType(expected.type, child.actual.type);
+                    org.objectweb.asm.Type descriptor = org.objectweb.asm.Type.getMethodType(actual.type, child.actual.type);
                     writer.invokeDynamic("plus", descriptor.getDescriptor(), DEF_BOOTSTRAP_HANDLE, DefBootstrap.UNARY_OPERATOR);
                 } 
             } else {
