+++ modules/lang-painless/src/test/java/org/elasticsearch/painless/DivisionTests.java
@@ -24,6 +24,11 @@ package org.elasticsearch.painless;
 public class DivisionTests extends ScriptTestCase {
 
     // TODO: byte,short,char
+    
+    public void testBasics() throws Exception {
+        assertEquals(2.25F / 1.5F, exec("return 2.25F / 1.5F;"));
+        assertEquals(0.5, exec("double x = 1; float y = 2; return x / y;"));
+    }
 
     public void testInt() throws Exception {
         assertEquals(1/1, exec("int x = 1; int y = 1; return x/y;"));
@@ -131,5 +136,203 @@ public class DivisionTests extends ScriptTestCase {
         expectScriptThrows(ArithmeticException.class, () -> {
             exec("return 1L/0L;");
         });
+    }
+    
+    public void testDef() {
+        assertEquals(1, exec("def x = (byte)2; def y = (byte)2; return x / y"));
+        assertEquals(1, exec("def x = (short)2; def y = (byte)2; return x / y"));
+        assertEquals(1, exec("def x = (char)2; def y = (byte)2; return x / y"));
+        assertEquals(1, exec("def x = (int)2; def y = (byte)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; def y = (byte)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; def y = (byte)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; def y = (byte)2; return x / y"));
+
+        assertEquals(1, exec("def x = (byte)2; def y = (short)2; return x / y"));
+        assertEquals(1, exec("def x = (short)2; def y = (short)2; return x / y"));
+        assertEquals(1, exec("def x = (char)2; def y = (short)2; return x / y"));
+        assertEquals(1, exec("def x = (int)2; def y = (short)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; def y = (short)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; def y = (short)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; def y = (short)2; return x / y"));
+
+        assertEquals(1, exec("def x = (byte)2; def y = (char)2; return x / y"));
+        assertEquals(1, exec("def x = (short)2; def y = (char)2; return x / y"));
+        assertEquals(1, exec("def x = (char)2; def y = (char)2; return x / y"));
+        assertEquals(1, exec("def x = (int)2; def y = (char)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; def y = (char)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; def y = (char)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; def y = (char)2; return x / y"));
+
+        assertEquals(1, exec("def x = (byte)2; def y = (int)2; return x / y"));
+        assertEquals(1, exec("def x = (short)2; def y = (int)2; return x / y"));
+        assertEquals(1, exec("def x = (char)2; def y = (int)2; return x / y"));
+        assertEquals(1, exec("def x = (int)2; def y = (int)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; def y = (int)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; def y = (int)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; def y = (int)2; return x / y"));
+
+        assertEquals(1L, exec("def x = (byte)2; def y = (long)2; return x / y"));
+        assertEquals(1L, exec("def x = (short)2; def y = (long)2; return x / y"));
+        assertEquals(1L, exec("def x = (char)2; def y = (long)2; return x / y"));
+        assertEquals(1L, exec("def x = (int)2; def y = (long)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; def y = (long)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; def y = (long)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; def y = (long)2; return x / y"));
+
+        assertEquals(1F, exec("def x = (byte)2; def y = (float)2; return x / y"));
+        assertEquals(1F, exec("def x = (short)2; def y = (float)2; return x / y"));
+        assertEquals(1F, exec("def x = (char)2; def y = (float)2; return x / y"));
+        assertEquals(1F, exec("def x = (int)2; def y = (float)2; return x / y"));
+        assertEquals(1F, exec("def x = (long)2; def y = (float)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; def y = (float)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; def y = (float)2; return x / y"));
+
+        assertEquals(1D, exec("def x = (byte)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (short)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (char)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (int)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (long)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (float)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; def y = (double)2; return x / y"));
+
+        assertEquals(1, exec("def x = (byte)2; def y = (byte)2; return x / y"));
+        assertEquals(1, exec("def x = (short)2; def y = (short)2; return x / y"));
+        assertEquals(1, exec("def x = (char)2; def y = (char)2; return x / y"));
+        assertEquals(1, exec("def x = (int)2; def y = (int)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; def y = (long)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; def y = (float)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; def y = (double)2; return x / y"));
+    }
+    
+    public void testDefTypedLHS() {
+        assertEquals(1, exec("byte x = (byte)2; def y = (byte)2; return x / y"));
+        assertEquals(1, exec("short x = (short)2; def y = (byte)2; return x / y"));
+        assertEquals(1, exec("char x = (char)2; def y = (byte)2; return x / y"));
+        assertEquals(1, exec("int x = (int)2; def y = (byte)2; return x / y"));
+        assertEquals(1L, exec("long x = (long)2; def y = (byte)2; return x / y"));
+        assertEquals(1F, exec("float x = (float)2; def y = (byte)2; return x / y"));
+        assertEquals(1D, exec("double x = (double)2; def y = (byte)2; return x / y"));
+
+        assertEquals(1, exec("byte x = (byte)2; def y = (short)2; return x / y"));
+        assertEquals(1, exec("short x = (short)2; def y = (short)2; return x / y"));
+        assertEquals(1, exec("char x = (char)2; def y = (short)2; return x / y"));
+        assertEquals(1, exec("int x = (int)2; def y = (short)2; return x / y"));
+        assertEquals(1L, exec("long x = (long)2; def y = (short)2; return x / y"));
+        assertEquals(1F, exec("float x = (float)2; def y = (short)2; return x / y"));
+        assertEquals(1D, exec("double x = (double)2; def y = (short)2; return x / y"));
+
+        assertEquals(1, exec("byte x = (byte)2; def y = (char)2; return x / y"));
+        assertEquals(1, exec("short x = (short)2; def y = (char)2; return x / y"));
+        assertEquals(1, exec("char x = (char)2; def y = (char)2; return x / y"));
+        assertEquals(1, exec("int x = (int)2; def y = (char)2; return x / y"));
+        assertEquals(1L, exec("long x = (long)2; def y = (char)2; return x / y"));
+        assertEquals(1F, exec("float x = (float)2; def y = (char)2; return x / y"));
+        assertEquals(1D, exec("double x = (double)2; def y = (char)2; return x / y"));
+
+        assertEquals(1, exec("byte x = (byte)2; def y = (int)2; return x / y"));
+        assertEquals(1, exec("short x = (short)2; def y = (int)2; return x / y"));
+        assertEquals(1, exec("char x = (char)2; def y = (int)2; return x / y"));
+        assertEquals(1, exec("int x = (int)2; def y = (int)2; return x / y"));
+        assertEquals(1L, exec("long x = (long)2; def y = (int)2; return x / y"));
+        assertEquals(1F, exec("float x = (float)2; def y = (int)2; return x / y"));
+        assertEquals(1D, exec("double x = (double)2; def y = (int)2; return x / y"));
+
+        assertEquals(1L, exec("byte x = (byte)2; def y = (long)2; return x / y"));
+        assertEquals(1L, exec("short x = (short)2; def y = (long)2; return x / y"));
+        assertEquals(1L, exec("char x = (char)2; def y = (long)2; return x / y"));
+        assertEquals(1L, exec("int x = (int)2; def y = (long)2; return x / y"));
+        assertEquals(1L, exec("long x = (long)2; def y = (long)2; return x / y"));
+        assertEquals(1F, exec("float x = (float)2; def y = (long)2; return x / y"));
+        assertEquals(1D, exec("double x = (double)2; def y = (long)2; return x / y"));
+
+        assertEquals(1F, exec("byte x = (byte)2; def y = (float)2; return x / y"));
+        assertEquals(1F, exec("short x = (short)2; def y = (float)2; return x / y"));
+        assertEquals(1F, exec("char x = (char)2; def y = (float)2; return x / y"));
+        assertEquals(1F, exec("int x = (int)2; def y = (float)2; return x / y"));
+        assertEquals(1F, exec("long x = (long)2; def y = (float)2; return x / y"));
+        assertEquals(1F, exec("float x = (float)2; def y = (float)2; return x / y"));
+        assertEquals(1D, exec("double x = (double)2; def y = (float)2; return x / y"));
+
+        assertEquals(1D, exec("byte x = (byte)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("short x = (short)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("char x = (char)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("int x = (int)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("long x = (long)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("float x = (float)2; def y = (double)2; return x / y"));
+        assertEquals(1D, exec("double x = (double)2; def y = (double)2; return x / y"));
+
+        assertEquals(1, exec("byte x = (byte)2; def y = (byte)2; return x / y"));
+        assertEquals(1, exec("short x = (short)2; def y = (short)2; return x / y"));
+        assertEquals(1, exec("char x = (char)2; def y = (char)2; return x / y"));
+        assertEquals(1, exec("int x = (int)2; def y = (int)2; return x / y"));
+        assertEquals(1L, exec("long x = (long)2; def y = (long)2; return x / y"));
+        assertEquals(1F, exec("float x = (float)2; def y = (float)2; return x / y"));
+        assertEquals(1D, exec("double x = (double)2; def y = (double)2; return x / y"));
+    }
+    
+    public void testDefTypedRHS() {
+        assertEquals(1, exec("def x = (byte)2; byte y = (byte)2; return x / y"));
+        assertEquals(1, exec("def x = (short)2; byte y = (byte)2; return x / y"));
+        assertEquals(1, exec("def x = (char)2; byte y = (byte)2; return x / y"));
+        assertEquals(1, exec("def x = (int)2; byte y = (byte)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; byte y = (byte)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; byte y = (byte)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; byte y = (byte)2; return x / y"));
+
+        assertEquals(1, exec("def x = (byte)2; short y = (short)2; return x / y"));
+        assertEquals(1, exec("def x = (short)2; short y = (short)2; return x / y"));
+        assertEquals(1, exec("def x = (char)2; short y = (short)2; return x / y"));
+        assertEquals(1, exec("def x = (int)2; short y = (short)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; short y = (short)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; short y = (short)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; short y = (short)2; return x / y"));
+
+        assertEquals(1, exec("def x = (byte)2; char y = (char)2; return x / y"));
+        assertEquals(1, exec("def x = (short)2; char y = (char)2; return x / y"));
+        assertEquals(1, exec("def x = (char)2; char y = (char)2; return x / y"));
+        assertEquals(1, exec("def x = (int)2; char y = (char)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; char y = (char)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; char y = (char)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; char y = (char)2; return x / y"));
+
+        assertEquals(1, exec("def x = (byte)2; int y = (int)2; return x / y"));
+        assertEquals(1, exec("def x = (short)2; int y = (int)2; return x / y"));
+        assertEquals(1, exec("def x = (char)2; int y = (int)2; return x / y"));
+        assertEquals(1, exec("def x = (int)2; int y = (int)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; int y = (int)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; int y = (int)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; int y = (int)2; return x / y"));
+
+        assertEquals(1L, exec("def x = (byte)2; long y = (long)2; return x / y"));
+        assertEquals(1L, exec("def x = (short)2; long y = (long)2; return x / y"));
+        assertEquals(1L, exec("def x = (char)2; long y = (long)2; return x / y"));
+        assertEquals(1L, exec("def x = (int)2; long y = (long)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; long y = (long)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; long y = (long)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; long y = (long)2; return x / y"));
+
+        assertEquals(1F, exec("def x = (byte)2; float y = (float)2; return x / y"));
+        assertEquals(1F, exec("def x = (short)2; float y = (float)2; return x / y"));
+        assertEquals(1F, exec("def x = (char)2; float y = (float)2; return x / y"));
+        assertEquals(1F, exec("def x = (int)2; float y = (float)2; return x / y"));
+        assertEquals(1F, exec("def x = (long)2; float y = (float)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; float y = (float)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; float y = (float)2; return x / y"));
+
+        assertEquals(1D, exec("def x = (byte)2; double y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (short)2; double y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (char)2; double y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (int)2; double y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (long)2; double y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (float)2; double y = (double)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; double y = (double)2; return x / y"));
+
+        assertEquals(1, exec("def x = (byte)2; byte y = (byte)2; return x / y"));
+        assertEquals(1, exec("def x = (short)2; short y = (short)2; return x / y"));
+        assertEquals(1, exec("def x = (char)2; char y = (char)2; return x / y"));
+        assertEquals(1, exec("def x = (int)2; int y = (int)2; return x / y"));
+        assertEquals(1L, exec("def x = (long)2; long y = (long)2; return x / y"));
+        assertEquals(1F, exec("def x = (float)2; float y = (float)2; return x / y"));
+        assertEquals(1D, exec("def x = (double)2; double y = (double)2; return x / y"));
     }
 }
