+++ modules/lang-painless/src/test/java/org/elasticsearch/painless/PromotionTests.java
@@ -19,93 +19,7 @@
 
 package org.elasticsearch.painless;
 
-/**
- * Tests binary operators across different types
- */
-// TODO: NaN/Inf/overflow/...
-public class BinaryOperatorTests extends ScriptTestCase {
-
-    // TODO: move to per-type tests and test for each type
-    public void testBasics() {
-        assertEquals(2.25F / 1.5F, exec("return 2.25F / 1.5F;"));
-        assertEquals(2.25F % 1.5F, exec("return 2.25F % 1.5F;"));
-        assertEquals(2 - 1, exec("return 2 - 1;"));
-        assertEquals(1 << 2, exec("return 1 << 2;"));
-        assertEquals(4 >> 2, exec("return 4 >> 2;"));
-        assertEquals(-1 >>> 29, exec("return -1 >>> 29;"));
-        assertEquals(5 & 3, exec("return 5 & 3;"));
-        assertEquals(5 & 3L, exec("return 5 & 3L;"));
-        assertEquals(5L & 3, exec("return 5L & 3;"));
-        assertEquals(5 | 3, exec("return 5 | 3;"));
-        assertEquals(5L | 3, exec("return 5L | 3;"));
-        assertEquals(5 | 3L, exec("return 5 | 3L;"));
-        assertEquals(9 ^ 3, exec("return 9 ^ 3;"));
-        assertEquals(9L ^ 3, exec("return 9L ^ 3;"));
-        assertEquals(9 ^ 3L, exec("return 9 ^ 3L;"));
-    }
-
-    public void testLongShifts() {
-        assertEquals(1L << 2, exec("long x = 1L; int y = 2; return x << y;"));
-        assertEquals(1 << 2L, exec("int x = 1; long y = 2L; return x << y;"));
-        assertEquals(4 >> 2L, exec("int x = 4; long y = 2L; return x >> y;"));
-        assertEquals(4L >> 2, exec("long x = 4L; int y = 2; return x >> y;"));
-        assertEquals(-1L >>> 29, exec("long x = -1L; int y = 29; return x >>> y;"));
-        assertEquals(-1 >>> 29L, exec("int x = -1; long y = 29L; return x >>> y;"));
-    }
-
-    public void testLongShiftsConst() {
-        assertEquals(1L << 2, exec("return 1L << 2;"));
-        assertEquals(1 << 2L, exec("return 1 << 2L;"));
-        assertEquals(4 >> 2L, exec("return 4 >> 2L;"));
-        assertEquals(4L >> 2, exec("return 4L >> 2;"));
-        assertEquals(-1L >>> 29, exec("return -1L >>> 29;"));
-        assertEquals(-1 >>> 29L, exec("return -1 >>> 29L;"));
-    }
-    
-    public void testBogusShifts() {
-        expectScriptThrows(ClassCastException.class, ()-> {
-            exec("long x = 1L; float y = 2; return x << y;");
-        });
-        expectScriptThrows(ClassCastException.class, ()-> {
-            exec("int x = 1; double y = 2L; return x << y;");
-        });
-        expectScriptThrows(ClassCastException.class, ()-> {
-            exec("float x = 1F; int y = 2; return x << y;");
-        });
-        expectScriptThrows(ClassCastException.class, ()-> {
-            exec("double x = 1D; int y = 2L; return x << y;");
-        });
-    }
-
-    public void testBogusShiftsConst() {
-        expectScriptThrows(ClassCastException.class, ()-> {
-            exec("return 1L << 2F;");
-        });
-        expectScriptThrows(ClassCastException.class, ()-> {
-            exec("return 1L << 2.0;");
-        });
-        expectScriptThrows(ClassCastException.class, ()-> {
-            exec("return 1F << 2;");
-        });
-        expectScriptThrows(ClassCastException.class, ()-> {
-            exec("return 1D << 2L");
-        });
-    }
-
-    public void testMixedTypes() {
-        assertEquals(8, exec("int x = 4; char y = 2; return x*y;"));
-        assertEquals(0.5, exec("double x = 1; float y = 2; return x / y;"));
-        assertEquals(1, exec("int x = 3; int y = 2; return x % y;"));
-        assertEquals(3.0, exec("double x = 1; byte y = 2; return x + y;"));
-        assertEquals(-1, exec("int x = 1; char y = 2; return x - y;"));
-        assertEquals(4, exec("int x = 1; char y = 2; return x << y;"));
-        assertEquals(-1, exec("int x = -1; char y = 29; return x >> y;"));
-        assertEquals(3, exec("int x = -1; char y = 30; return x >>> y;"));
-        assertEquals(1L, exec("int x = 5; long y = 3; return x & y;"));
-        assertEquals(7, exec("short x = 5; byte y = 3; return x | y;"));
-        assertEquals(10, exec("short x = 9; char y = 3; return x ^ y;"));
-    }
-
+public class PromotionTests extends ScriptTestCase {
     public void testBinaryPromotion() throws Exception {
         // byte/byte
         assertEquals((byte)1 + (byte)1, exec("byte x = 1; byte y = 1; return x+y;"));
