+++ modules/lang-painless/src/test/java/org/elasticsearch/painless/RemainderTests.java
@@ -24,6 +24,11 @@ package org.elasticsearch.painless;
 public class RemainderTests extends ScriptTestCase {
 
     // TODO: byte,short,char
+    
+    public void testBasics() throws Exception {
+        assertEquals(2.25F % 1.5F, exec("return 2.25F % 1.5F;"));
+        assertEquals(1, exec("int x = 3; int y = 2; return x % y;"));
+    }
 
     public void testInt() throws Exception {
         assertEquals(1%1, exec("int x = 1; int y = 1; return x%y;"));
@@ -131,5 +136,203 @@ public class RemainderTests extends ScriptTestCase {
         expectScriptThrows(ArithmeticException.class, () -> {
             exec("return 1L%0L;");
         });
+    }
+    
+    public void testDef() {
+        assertEquals(0, exec("def x = (byte)2; def y = (byte)2; return x % y"));
+        assertEquals(0, exec("def x = (short)2; def y = (byte)2; return x % y"));
+        assertEquals(0, exec("def x = (char)2; def y = (byte)2; return x % y"));
+        assertEquals(0, exec("def x = (int)2; def y = (byte)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; def y = (byte)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; def y = (byte)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; def y = (byte)2; return x % y"));
+
+        assertEquals(0, exec("def x = (byte)2; def y = (short)2; return x % y"));
+        assertEquals(0, exec("def x = (short)2; def y = (short)2; return x % y"));
+        assertEquals(0, exec("def x = (char)2; def y = (short)2; return x % y"));
+        assertEquals(0, exec("def x = (int)2; def y = (short)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; def y = (short)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; def y = (short)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; def y = (short)2; return x % y"));
+
+        assertEquals(0, exec("def x = (byte)2; def y = (char)2; return x % y"));
+        assertEquals(0, exec("def x = (short)2; def y = (char)2; return x % y"));
+        assertEquals(0, exec("def x = (char)2; def y = (char)2; return x % y"));
+        assertEquals(0, exec("def x = (int)2; def y = (char)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; def y = (char)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; def y = (char)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; def y = (char)2; return x % y"));
+
+        assertEquals(0, exec("def x = (byte)2; def y = (int)2; return x % y"));
+        assertEquals(0, exec("def x = (short)2; def y = (int)2; return x % y"));
+        assertEquals(0, exec("def x = (char)2; def y = (int)2; return x % y"));
+        assertEquals(0, exec("def x = (int)2; def y = (int)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; def y = (int)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; def y = (int)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; def y = (int)2; return x % y"));
+
+        assertEquals(0L, exec("def x = (byte)2; def y = (long)2; return x % y"));
+        assertEquals(0L, exec("def x = (short)2; def y = (long)2; return x % y"));
+        assertEquals(0L, exec("def x = (char)2; def y = (long)2; return x % y"));
+        assertEquals(0L, exec("def x = (int)2; def y = (long)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; def y = (long)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; def y = (long)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; def y = (long)2; return x % y"));
+
+        assertEquals(0F, exec("def x = (byte)2; def y = (float)2; return x % y"));
+        assertEquals(0F, exec("def x = (short)2; def y = (float)2; return x % y"));
+        assertEquals(0F, exec("def x = (char)2; def y = (float)2; return x % y"));
+        assertEquals(0F, exec("def x = (int)2; def y = (float)2; return x % y"));
+        assertEquals(0F, exec("def x = (long)2; def y = (float)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; def y = (float)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; def y = (float)2; return x % y"));
+
+        assertEquals(0D, exec("def x = (byte)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (short)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (char)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (int)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (long)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (float)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; def y = (double)2; return x % y"));
+
+        assertEquals(0, exec("def x = (byte)2; def y = (byte)2; return x % y"));
+        assertEquals(0, exec("def x = (short)2; def y = (short)2; return x % y"));
+        assertEquals(0, exec("def x = (char)2; def y = (char)2; return x % y"));
+        assertEquals(0, exec("def x = (int)2; def y = (int)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; def y = (long)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; def y = (float)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; def y = (double)2; return x % y"));
+    }
+    
+    public void testDefTypedLHS() {
+        assertEquals(0, exec("byte x = (byte)2; def y = (byte)2; return x % y"));
+        assertEquals(0, exec("short x = (short)2; def y = (byte)2; return x % y"));
+        assertEquals(0, exec("char x = (char)2; def y = (byte)2; return x % y"));
+        assertEquals(0, exec("int x = (int)2; def y = (byte)2; return x % y"));
+        assertEquals(0L, exec("long x = (long)2; def y = (byte)2; return x % y"));
+        assertEquals(0F, exec("float x = (float)2; def y = (byte)2; return x % y"));
+        assertEquals(0D, exec("double x = (double)2; def y = (byte)2; return x % y"));
+
+        assertEquals(0, exec("byte x = (byte)2; def y = (short)2; return x % y"));
+        assertEquals(0, exec("short x = (short)2; def y = (short)2; return x % y"));
+        assertEquals(0, exec("char x = (char)2; def y = (short)2; return x % y"));
+        assertEquals(0, exec("int x = (int)2; def y = (short)2; return x % y"));
+        assertEquals(0L, exec("long x = (long)2; def y = (short)2; return x % y"));
+        assertEquals(0F, exec("float x = (float)2; def y = (short)2; return x % y"));
+        assertEquals(0D, exec("double x = (double)2; def y = (short)2; return x % y"));
+
+        assertEquals(0, exec("byte x = (byte)2; def y = (char)2; return x % y"));
+        assertEquals(0, exec("short x = (short)2; def y = (char)2; return x % y"));
+        assertEquals(0, exec("char x = (char)2; def y = (char)2; return x % y"));
+        assertEquals(0, exec("int x = (int)2; def y = (char)2; return x % y"));
+        assertEquals(0L, exec("long x = (long)2; def y = (char)2; return x % y"));
+        assertEquals(0F, exec("float x = (float)2; def y = (char)2; return x % y"));
+        assertEquals(0D, exec("double x = (double)2; def y = (char)2; return x % y"));
+
+        assertEquals(0, exec("byte x = (byte)2; def y = (int)2; return x % y"));
+        assertEquals(0, exec("short x = (short)2; def y = (int)2; return x % y"));
+        assertEquals(0, exec("char x = (char)2; def y = (int)2; return x % y"));
+        assertEquals(0, exec("int x = (int)2; def y = (int)2; return x % y"));
+        assertEquals(0L, exec("long x = (long)2; def y = (int)2; return x % y"));
+        assertEquals(0F, exec("float x = (float)2; def y = (int)2; return x % y"));
+        assertEquals(0D, exec("double x = (double)2; def y = (int)2; return x % y"));
+
+        assertEquals(0L, exec("byte x = (byte)2; def y = (long)2; return x % y"));
+        assertEquals(0L, exec("short x = (short)2; def y = (long)2; return x % y"));
+        assertEquals(0L, exec("char x = (char)2; def y = (long)2; return x % y"));
+        assertEquals(0L, exec("int x = (int)2; def y = (long)2; return x % y"));
+        assertEquals(0L, exec("long x = (long)2; def y = (long)2; return x % y"));
+        assertEquals(0F, exec("float x = (float)2; def y = (long)2; return x % y"));
+        assertEquals(0D, exec("double x = (double)2; def y = (long)2; return x % y"));
+
+        assertEquals(0F, exec("byte x = (byte)2; def y = (float)2; return x % y"));
+        assertEquals(0F, exec("short x = (short)2; def y = (float)2; return x % y"));
+        assertEquals(0F, exec("char x = (char)2; def y = (float)2; return x % y"));
+        assertEquals(0F, exec("int x = (int)2; def y = (float)2; return x % y"));
+        assertEquals(0F, exec("long x = (long)2; def y = (float)2; return x % y"));
+        assertEquals(0F, exec("float x = (float)2; def y = (float)2; return x % y"));
+        assertEquals(0D, exec("double x = (double)2; def y = (float)2; return x % y"));
+
+        assertEquals(0D, exec("byte x = (byte)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("short x = (short)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("char x = (char)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("int x = (int)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("long x = (long)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("float x = (float)2; def y = (double)2; return x % y"));
+        assertEquals(0D, exec("double x = (double)2; def y = (double)2; return x % y"));
+
+        assertEquals(0, exec("byte x = (byte)2; def y = (byte)2; return x % y"));
+        assertEquals(0, exec("short x = (short)2; def y = (short)2; return x % y"));
+        assertEquals(0, exec("char x = (char)2; def y = (char)2; return x % y"));
+        assertEquals(0, exec("int x = (int)2; def y = (int)2; return x % y"));
+        assertEquals(0L, exec("long x = (long)2; def y = (long)2; return x % y"));
+        assertEquals(0F, exec("float x = (float)2; def y = (float)2; return x % y"));
+        assertEquals(0D, exec("double x = (double)2; def y = (double)2; return x % y"));
+    }
+
+    public void testDefTypedRHS() {
+        assertEquals(0, exec("def x = (byte)2; byte y = (byte)2; return x % y"));
+        assertEquals(0, exec("def x = (short)2; byte y = (byte)2; return x % y"));
+        assertEquals(0, exec("def x = (char)2; byte y = (byte)2; return x % y"));
+        assertEquals(0, exec("def x = (int)2; byte y = (byte)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; byte y = (byte)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; byte y = (byte)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; byte y = (byte)2; return x % y"));
+
+        assertEquals(0, exec("def x = (byte)2; short y = (short)2; return x % y"));
+        assertEquals(0, exec("def x = (short)2; short y = (short)2; return x % y"));
+        assertEquals(0, exec("def x = (char)2; short y = (short)2; return x % y"));
+        assertEquals(0, exec("def x = (int)2; short y = (short)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; short y = (short)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; short y = (short)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; short y = (short)2; return x % y"));
+
+        assertEquals(0, exec("def x = (byte)2; char y = (char)2; return x % y"));
+        assertEquals(0, exec("def x = (short)2; char y = (char)2; return x % y"));
+        assertEquals(0, exec("def x = (char)2; char y = (char)2; return x % y"));
+        assertEquals(0, exec("def x = (int)2; char y = (char)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; char y = (char)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; char y = (char)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; char y = (char)2; return x % y"));
+
+        assertEquals(0, exec("def x = (byte)2; int y = (int)2; return x % y"));
+        assertEquals(0, exec("def x = (short)2; int y = (int)2; return x % y"));
+        assertEquals(0, exec("def x = (char)2; int y = (int)2; return x % y"));
+        assertEquals(0, exec("def x = (int)2; int y = (int)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; int y = (int)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; int y = (int)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; int y = (int)2; return x % y"));
+
+        assertEquals(0L, exec("def x = (byte)2; long y = (long)2; return x % y"));
+        assertEquals(0L, exec("def x = (short)2; long y = (long)2; return x % y"));
+        assertEquals(0L, exec("def x = (char)2; long y = (long)2; return x % y"));
+        assertEquals(0L, exec("def x = (int)2; long y = (long)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; long y = (long)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; long y = (long)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; long y = (long)2; return x % y"));
+
+        assertEquals(0F, exec("def x = (byte)2; float y = (float)2; return x % y"));
+        assertEquals(0F, exec("def x = (short)2; float y = (float)2; return x % y"));
+        assertEquals(0F, exec("def x = (char)2; float y = (float)2; return x % y"));
+        assertEquals(0F, exec("def x = (int)2; float y = (float)2; return x % y"));
+        assertEquals(0F, exec("def x = (long)2; float y = (float)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; float y = (float)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; float y = (float)2; return x % y"));
+
+        assertEquals(0D, exec("def x = (byte)2; double y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (short)2; double y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (char)2; double y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (int)2; double y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (long)2; double y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (float)2; double y = (double)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; double y = (double)2; return x % y"));
+
+        assertEquals(0, exec("def x = (byte)2; byte y = (byte)2; return x % y"));
+        assertEquals(0, exec("def x = (short)2; short y = (short)2; return x % y"));
+        assertEquals(0, exec("def x = (char)2; char y = (char)2; return x % y"));
+        assertEquals(0, exec("def x = (int)2; int y = (int)2; return x % y"));
+        assertEquals(0L, exec("def x = (long)2; long y = (long)2; return x % y"));
+        assertEquals(0F, exec("def x = (float)2; float y = (float)2; return x % y"));
+        assertEquals(0D, exec("def x = (double)2; double y = (double)2; return x % y"));
     }
 }
