+++ modules/lang-painless/src/test/java/org/elasticsearch/painless/UnaryTests.java
@@ -49,4 +49,53 @@ public class UnaryTests extends ScriptTestCase {
         assertEquals(-1.0F, exec("float x = -1F; return +x"));
         assertEquals(-1.0, exec("double x = -1.0; return +x"));
     }
+    
+    public void testDefNot() {
+        assertEquals(~1, exec("def x = (byte)1; return ~x"));
+        assertEquals(~1, exec("def x = (short)1; return ~x"));
+        assertEquals(~1, exec("def x = (char)1; return ~x"));
+        assertEquals(~1, exec("def x = 1; return ~x"));
+        assertEquals(~1L, exec("def x = 1L; return ~x"));
+    }
+    
+    public void testDefNotTypedRet() {
+        assertEquals((double)~1, exec("def x = (byte)1; double y = ~x; return y;"));
+        assertEquals((float)~1, exec("def x = (short)1; float y = ~x; return y;"));
+        assertEquals((long)~1, exec("def x = (char)1; long y = ~x; return y;"));
+        assertEquals(~1, exec("def x = 1; int y = ~x; return y;"));
+    }
+
+    public void testDefNeg() {
+        assertEquals(-1, exec("def x = (byte)1; return -x"));
+        assertEquals(-1, exec("def x = (short)1; return -x"));
+        assertEquals(-1, exec("def x = (char)1; return -x"));
+        assertEquals(-1, exec("def x = 1; return -x"));
+        assertEquals(-1L, exec("def x = 1L; return -x"));
+        assertEquals(-1.0F, exec("def x = 1F; return -x"));
+        assertEquals(-1.0, exec("def x = 1.0; return -x"));
+    }
+    
+    public void testDefNegTypedRet() {
+        assertEquals((double)-1, exec("def x = (byte)1; double y = -x; return y;"));
+        assertEquals((float)-1, exec("def x = (short)1; float y = -x; return y;"));
+        assertEquals((long)-1, exec("def x = (char)1; long y = -x; return y;"));
+        assertEquals(-1, exec("def x = 1; int y = -x; return y;"));
+    }
+    
+    public void testDefPlus() {
+        assertEquals(-1, exec("def x = (byte)-1; return +x"));
+        assertEquals(-1, exec("def x = (short)-1; return +x"));
+        assertEquals(65535, exec("def x = (char)-1; return +x"));
+        assertEquals(-1, exec("def x = -1; return +x"));
+        assertEquals(-1L, exec("def x = -1L; return +x"));
+        assertEquals(-1.0F, exec("def x = -1F; return +x"));
+        assertEquals(-1.0D, exec("def x = -1.0; return +x"));
+    }
+    
+    public void testDefPlusTypedRet() {
+        assertEquals((double)-1, exec("def x = (byte)-1; double y = +x; return y;"));
+        assertEquals((float)-1, exec("def x = (short)-1; float y = +x; return y;"));
+        assertEquals((long)65535, exec("def x = (char)-1; long y = +x; return y;"));
+        assertEquals(-1, exec("def x = -1; int y = +x; return y;"));
+    }
 }
